# Make Slicer interface more generic