… object/slicer/slicer_test.go
… pool/object.go
… pool/pool.go
… pool/pool_aio_test.go
… pool/pool_test.go