closes #442