@smallhive 2023-07-12
> im a bit confused that we require to pass `object.Object`: imo this can complicate the code of a simple data uploading applications.
> 
> how about provide both raw interface (current) accepting `object.Object` and the sugared (previous) one requiring min input? @smallhive @carpawell @roman-khimov

I consider, it doesn't make sense. User has to pass all required parameters for the header object (signer neofscrypto.Signer, cnr cid.ID, owner user.ID) and we create it inside.
Passing object.Object gives a flexible way to configure the header. In this case, the user creates the header itself with the same data.

For comparison, we manipulate the same data, but with different ways:

Before
```
slicer := new(signer neofscrypto.Signer, cnr cid.ID, owner user.ID, w ObjectWriter, opts Options)
slicer.Slice(data io.Reader, attributes ...string)
```
After
```
var hdr object.Object
header.SetContainerID(cnr)
header.SetOwnerID(&owner)
header.SetAttributes(...)
header.///

slicer := new(w ObjectWriter)
slicer.Slice(header object.Object, data io.Reader, signer neofscrypto.Signer, opts Options)
```
Yes, the code longer for header creation, but it is deadly simple