@cthulhu-rider 2023-07-13
> header.///

that's what exactly bothers me. How user will undestand which fields must be explicitly set and which are optional or event must not be set?

accepting `cid.ID`, `user.ID`, `neofscrypto.Signer`, string attibutes and `io.Reader` covers most of the external uploader's needs imo. And accepting header structure (`object.Object` as done in this PR) is good for system needs and small part of deeply understanding users. Simple splitted input can call header one within package (from simple to flexible).

i can also suggest to implement customization through optional header construction callback that is going to be called during object formation if set:
```go
var opts Options
opts.SetCustomHeaderConstructor(func(obj *object.Object) {
  obj.SetType(object.Tombstone)
  obj.SetWhatever()
})

slicer := new(signer neofscrypto.Signer, cnr cid.ID, owner user.ID, w ObjectWriter, opts Options)
slicer.SliceData(bytes.NewReader(tombstone.Marshal()))
```