@smallhive 2023-07-13
> > header.///
> 
> that's what exactly bothers me. How user will undestand which fields must be explicitly set and which are optional or even must not be set?

It's not a problem. We can create a helper function inside the `slicer` package. For instance:
```go
func PrepareHeader(cnr cid.ID, owner oid.ID) object.Object
```
Also, we will mention it inside the Slicer constructor/Slice method. Any simple upload application will manage to call the func for header construction. This will be a good starting point for them, to touch our product. But, as practice shows, latter they would want to dive deeper. In this case they shouldn't rewrite a lot their code, because PrepareHeader is just 2 rows function.

> accepting `cid.ID`, `user.ID`, `neofscrypto.Signer`, string attibutes and `io.Reader` covers most of the external uploader's needs imo. And accepting header structure (`object.Object` as done in this PR) is good for system needs and small part of deeply understanding users. Simple splitted input can call header one within package (from simple to flexible).

I don't agree with strings attributes. Regular client and node already use object.Attribute and have to reconvert existing attributes to string. Me and @carpawell faced the situation.

> i can also suggest to implement customization through optional header construction callback that is going to be called during object formation if set:
> 
> ```go
> var opts Options
> opts.SetCustomHeaderConstructor(func(obj *object.Object) {
>   obj.SetType(object.Tombstone)
>   obj.SetWhatever()
> })
> slicer := new(signer neofscrypto.Signer, cnr cid.ID, owner user.ID, w ObjectWriter, opts Options)
> slicer.SliceData(bytes.NewReader(tombstone.Marshal()))
> ```

One more option var :smiley: Passing a fully prepared header eliminates all misunderstandings. This callback, we would 
 only call for the header, because all other parts are `TypeRegular`. Thus, such a callback looks excessive.

> P.S. me personally completely okay with raw API, but I'm not sure if it will be convenient and understandable for simple data upload applications. I'd provide helpers that transform what they have to required stuff like `hdr object.Object`.

Even the current SDK interface is quite clear and understandable. The slicer improves these factors, but I tend to think it shouldn't make such far abstraction from SDK types