@cthulhu-rider 2023-07-17
@smallhive i understand and agree with almost all ur points. Let me summarize my own

in my mind, `slicer` space is intended to simplify data upload. The min required set to upload the data into NeoFS is container ID, `Signer` and data itself encapsulated in `io.Reader` primitive. So, i'd like to provide API which allows to use this small input and perform uploading by min num of instructions (ideally - by 1 instruction). If we also want to support some more specific needs like tombstone formation, it's better to add specific API it next to the simple one, implement the simple API through the specific one and not force the user to understand the specifics (if he wants to, he will figure it out and use).

to sum it up,
```
s := slicer.New(signer, objectWriter, opts)
err := slicer.Slice(slicer.RequiredFields(cnr, owner), strings.NewReader("Hello, world!"))
```
looks pretty good to me. If this is the approach u suggested, i agree with it