@roman-khimov 2023-07-20
Maybe
```
s/ObjectPutter/ObjectClient/
s/NetworkedWriter/NetworkedClient/
s/Write/Put/
s/InitWrite/InitPut/
```
For better names and consistency with `client`.