@smallhive 2023-07-24
According to [the comment](https://github.com/nspcc-dev/neofs-sdk-go/pull/466#pullrequestreview-1537832050), completely rewrite the solution