@smallhive 2023-07-28
Updated commit message benchmark statistic, also final approach: master vs current branch

```
                                                 │   old.txt    │               new.txt               │
                                                 │    sec/op    │    sec/op     vs base               │
SliceDataIntoObjects/slice_1-128/reader-16         113.39µ ± 0%   62.77µ ±  0%  -44.64% (p=0.002 n=6)
SliceDataIntoObjects/slice_1-128/writer-16         117.84µ ± 1%   65.82µ ±  0%  -44.14% (p=0.002 n=6)
SliceDataIntoObjects/slice_4-128/reader-16         116.26µ ± 1%   67.40µ ±  1%  -42.02% (p=0.002 n=6)
SliceDataIntoObjects/slice_4-128/writer-16         121.26µ ± 3%   70.42µ ±  2%  -41.92% (p=0.002 n=6)
SliceDataIntoObjects/slice_16-128/reader-16        111.88µ ± 8%   65.74µ ±  1%  -41.24% (p=0.002 n=6)
SliceDataIntoObjects/slice_16-128/writer-16        116.83µ ± 2%   68.16µ ±  1%  -41.66% (p=0.002 n=6)
SliceDataIntoObjects/slice_64-128/reader-16        113.87µ ± 2%   66.09µ ±  1%  -41.96% (p=0.002 n=6)
SliceDataIntoObjects/slice_64-128/writer-16        120.53µ ± 2%   77.10µ ±  7%  -36.04% (p=0.002 n=6)
SliceDataIntoObjects/slice_256-128/reader-16       118.88µ ± 3%   70.01µ ± 10%  -41.11% (p=0.002 n=6)
SliceDataIntoObjects/slice_256-128/writer-16        254.7µ ± 4%   234.0µ ±  1%   -8.14% (p=0.002 n=6)
SliceDataIntoObjects/slice_1024-128/reader-16      120.60µ ± 3%   65.18µ ±  6%  -45.95% (p=0.002 n=6)
SliceDataIntoObjects/slice_1024-128/writer-16       645.1µ ± 1%   570.0µ ±  5%  -11.65% (p=0.002 n=6)
SliceDataIntoObjects/slice_4096-128/reader-16      118.30µ ± 3%   69.01µ ±  3%  -41.67% (p=0.002 n=6)
SliceDataIntoObjects/slice_4096-128/writer-16       2.131m ± 5%   2.024m ±  3%   -5.03% (p=0.002 n=6)
SliceDataIntoObjects/slice_16384-128/reader-16     112.72µ ± 2%   64.07µ ±  0%  -43.16% (p=0.002 n=6)
SliceDataIntoObjects/slice_16384-128/writer-16      7.266m ± 1%   6.792m ±  8%        ~ (p=0.065 n=6)
SliceDataIntoObjects/slice_65536-128/reader-16     115.29µ ± 3%   69.42µ ±  1%  -39.79% (p=0.002 n=6)
SliceDataIntoObjects/slice_65536-128/writer-16      28.75m ± 1%   30.11m ±  1%   +4.74% (p=0.002 n=6)
SliceDataIntoObjects/slice_262144-128/reader-16    131.98µ ± 4%   76.69µ ±  3%  -41.89% (p=0.002 n=6)
SliceDataIntoObjects/slice_262144-128/writer-16     121.1m ± 2%   123.6m ±  4%   +2.07% (p=0.041 n=6)
SliceDataIntoObjects/slice_1048576-128/reader-16    218.6µ ± 8%   106.8µ ± 12%  -51.15% (p=0.002 n=6)
SliceDataIntoObjects/slice_1048576-128/writer-16    528.9m ± 6%   444.4m ±  1%  -15.96% (p=0.002 n=6)
geomean                                             480.8µ        322.7µ        -32.89%

                                                 │    old.txt    │               new.txt               │
                                                 │     B/op      │     B/op      vs base               │
SliceDataIntoObjects/slice_1-128/reader-16         15.629Ki ± 0%   9.547Ki ± 0%  -38.92% (p=0.002 n=6)
SliceDataIntoObjects/slice_1-128/writer-16         16.786Ki ± 0%   9.421Ki ± 0%  -43.88% (p=0.002 n=6)
SliceDataIntoObjects/slice_4-128/reader-16          16.50Ki ± 0%   11.30Ki ± 0%  -31.55% (p=0.002 n=6)
SliceDataIntoObjects/slice_4-128/writer-16          17.66Ki ± 0%   11.17Ki ± 0%  -36.74% (p=0.002 n=6)
SliceDataIntoObjects/slice_16-128/reader-16         14.00Ki ± 0%   10.55Ki ± 0%  -24.67% (p=0.002 n=6)
SliceDataIntoObjects/slice_16-128/writer-16         15.16Ki ± 0%   10.42Ki ± 0%  -31.25% (p=0.002 n=6)
SliceDataIntoObjects/slice_64-128/reader-16         15.63Ki ± 0%   10.55Ki ± 0%  -32.51% (p=0.002 n=6)
SliceDataIntoObjects/slice_64-128/writer-16         16.79Ki ± 0%   10.42Ki ± 0%  -37.91% (p=0.002 n=6)
SliceDataIntoObjects/slice_256-128/reader-16       13.630Ki ± 0%   9.204Ki ± 0%  -32.47% (p=0.002 n=6)
SliceDataIntoObjects/slice_256-128/writer-16        30.83Ki ± 0%   29.42Ki ± 0%   -4.57% (p=0.002 n=6)
SliceDataIntoObjects/slice_1024-128/reader-16      16.887Ki ± 0%   9.675Ki ± 0%  -42.71% (p=0.002 n=6)
SliceDataIntoObjects/slice_1024-128/writer-16       81.64Ki ± 0%   69.60Ki ± 0%  -14.75% (p=0.002 n=6)
SliceDataIntoObjects/slice_4096-128/reader-16       15.28Ki ± 0%   11.31Ki ± 0%  -25.95% (p=0.002 n=6)
SliceDataIntoObjects/slice_4096-128/writer-16       255.3Ki ± 0%   235.7Ki ± 0%   -7.70% (p=0.002 n=6)
SliceDataIntoObjects/slice_16384-128/reader-16     13.205Ki ± 0%   9.716Ki ± 0%  -26.42% (p=0.002 n=6)
SliceDataIntoObjects/slice_16384-128/writer-16      932.1Ki ± 0%   831.3Ki ± 0%  -10.81% (p=0.002 n=6)
SliceDataIntoObjects/slice_65536-128/reader-16      13.48Ki ± 0%   11.50Ki ± 0%  -14.71% (p=0.002 n=6)
SliceDataIntoObjects/slice_65536-128/writer-16      3.595Mi ± 0%   3.307Mi ± 0%   -8.01% (p=0.002 n=6)
SliceDataIntoObjects/slice_262144-128/reader-16     16.32Ki ± 1%   12.32Ki ± 0%  -24.56% (p=0.002 n=6)
SliceDataIntoObjects/slice_262144-128/writer-16     14.45Mi ± 0%   13.26Mi ± 0%   -8.18% (p=0.002 n=6)
SliceDataIntoObjects/slice_1048576-128/reader-16    24.55Ki ± 5%   15.16Ki ± 6%  -38.24% (p=0.002 n=6)
SliceDataIntoObjects/slice_1048576-128/writer-16    59.47Mi ± 0%   51.22Mi ± 0%  -13.88% (p=0.002 n=6)
geomean                                             61.32Ki        45.35Ki       -26.03%

                                                 │   old.txt   │              new.txt               │
                                                 │  allocs/op  │  allocs/op   vs base               │
SliceDataIntoObjects/slice_1-128/reader-16          235.0 ± 0%    149.0 ± 0%  -36.60% (p=0.002 n=6)
SliceDataIntoObjects/slice_1-128/writer-16          256.0 ± 0%    148.0 ± 0%  -42.19% (p=0.002 n=6)
SliceDataIntoObjects/slice_4-128/reader-16          247.0 ± 0%    174.0 ± 0%  -29.55% (p=0.002 n=6)
SliceDataIntoObjects/slice_4-128/writer-16          268.0 ± 0%    173.0 ± 0%  -35.45% (p=0.002 n=6)
SliceDataIntoObjects/slice_16-128/reader-16         211.0 ± 0%    163.0 ± 0%  -22.75% (p=0.002 n=6)
SliceDataIntoObjects/slice_16-128/writer-16         232.0 ± 0%    162.0 ± 0%  -30.17% (p=0.002 n=6)
SliceDataIntoObjects/slice_64-128/reader-16         235.0 ± 0%    163.0 ± 0%  -30.64% (p=0.002 n=6)
SliceDataIntoObjects/slice_64-128/writer-16         256.0 ± 0%    162.0 ± 0%  -36.72% (p=0.002 n=6)
SliceDataIntoObjects/slice_256-128/reader-16        204.0 ± 0%    144.0 ± 0%  -29.41% (p=0.002 n=6)
SliceDataIntoObjects/slice_256-128/writer-16        458.0 ± 0%    437.0 ± 0%   -4.59% (p=0.002 n=6)
SliceDataIntoObjects/slice_1024-128/reader-16       254.0 ± 0%    151.0 ± 0%  -40.55% (p=0.002 n=6)
SliceDataIntoObjects/slice_1024-128/writer-16      1.250k ± 0%   1.030k ± 0%  -17.60% (p=0.002 n=6)
SliceDataIntoObjects/slice_4096-128/reader-16       230.0 ± 0%    174.0 ± 0%  -24.35% (p=0.002 n=6)
SliceDataIntoObjects/slice_4096-128/writer-16      3.936k ± 0%   3.464k ± 0%  -11.99% (p=0.002 n=6)
SliceDataIntoObjects/slice_16384-128/reader-16      198.0 ± 0%    151.0 ± 0%  -23.74% (p=0.002 n=6)
SliceDataIntoObjects/slice_16384-128/writer-16     14.44k ± 0%   12.20k ± 0%  -15.53% (p=0.002 n=6)
SliceDataIntoObjects/slice_65536-128/reader-16      202.0 ± 0%    176.0 ± 0%  -12.87% (p=0.002 n=6)
SliceDataIntoObjects/slice_65536-128/writer-16     57.08k ± 0%   49.56k ± 0%  -13.17% (p=0.002 n=6)
SliceDataIntoObjects/slice_262144-128/reader-16     245.0 ± 0%    188.0 ± 1%  -23.27% (p=0.002 n=6)
SliceDataIntoObjects/slice_262144-128/writer-16    227.7k ± 0%   197.1k ± 0%  -13.44% (p=0.002 n=6)
SliceDataIntoObjects/slice_1048576-128/reader-16    360.5 ± 5%    230.5 ± 6%  -36.06% (p=0.002 n=6)
SliceDataIntoObjects/slice_1048576-128/writer-16   934.4k ± 0%   762.4k ± 0%  -18.40% (p=0.002 n=6)
geomean                                             928.4         690.1       -25.67%
```